fix for linux vs mac ping (received vs packets received)
